Unify how chartJsOptions are generated for PerfherderGraphContainer and RedashContainer
This will prevent further divergence in the code.
