+++ src/utils/chartJs/perfherderFormatter.js
@@ -1,39 +1,14 @@
+import generateOptions from './generateOptions';
+
 const dataToChartJSformat = data => data.map(({ datetime, value }) => ({
     x: datetime,
     y: value,
   }));
 
-const chartJsOptions = (reverse, scaleLabel) => ({
-  scales: {
-    xAxes: [
-      {
-        type: 'time',
-        time: {
-          displayFormats: {
-            hour: 'MMM D',
-          },
-        },
-      },
-    ],
-    yAxes: [
-      {
-        ticks: {
-          beginAtZero: true,
-          reverse,
-        },
-        scaleLabel: {
-          display: true,
-          labelString: scaleLabel,
-        },
-      },
-    ],
-  },
-});
-
 const generateChartJsOptions = (meta) => {
   const higherIsBetter = (meta.lower_is_better === false);
   const yLabel = higherIsBetter ? 'Score' : 'Execution time (ms)';
-  return chartJsOptions(higherIsBetter, yLabel);
+  return generateOptions(higherIsBetter, yLabel);
 };
 
 // This function combines Perfherder series and transforms it into ChartJS formatting
